This is a combination of 2 commits.
The first commit's message is:
Created a python sandbox example based on sandboxtemplate and sandboxclaim using the controller extensions
which can now create a family of python code execution sandboxes.

Due to issues testing the new Sandboxtemplate and claim extension I fixed 2 issues below.

The second commit's message is:

1. The sandbox-python-claim.yaml was using a field named templateRef to point to the SandboxTemplate.
In extensions/api/v1alpha1/sandboxclaim_types.go, I found that the corresponding field in the SandboxClaimSpec
struct was named sandboxTemplateRef.
To fix this, I modified the Go code to match the YAML -  I renamed the TemplateRef field within the SandboxClaimSpec struct and updated its JSON tag from sandboxTemplateRef to templateRef.

2. In k8s/deployment.yaml I added:
   containers:
              - name: agent-sandbox-controller
                image: ko://sigs.k8s.io/agent-sandbox/cmd/agent-sandbox-controller
                args:
                - --extensions
So it starts with the extensions enabled and makes it recognize and process the SandboxClaim resource to create the Sandbox pod.